+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## [v0.7.1](https://github.com/VictoriaMetrics/grafana-datasource/releases/tag/v0.7.1)
 
+* FEATURE: add variable type selector to optimize API usage, favoring [/api/v1/labels](https://docs.victoriametrics.com/url-examples/#apiv1labels) and [/api/v1/label/.../values](https://docs.victoriametrics.com/url-examples/#apiv1labelvalues) over [/api/v1/series](https://docs.victoriametrics.com/url-examples/#apiv1series). See [this issue](https://github.com/VictoriaMetrics/grafana-datasource/issues/144)
+
 * BUGFIX: correct parsing the Annotations queries when template variables are used. See [this issue](https://github.com/VictoriaMetrics/grafana-datasource/issues/146)
 
 ## [v0.7.0](https://github.com/VictoriaMetrics/grafana-datasource/releases/tag/v0.7.0)